Move service loader to `loader.go`
